added styling

+++ budget_app/models.py
@@ -5,7 +5,7 @@ from django.contrib.auth.models import User
 class BudgetItem(models.Model):
     title = models.CharField(max_length=100)
     description = models.CharField(max_length=100, blank=True)
-    amount = models.IntegerField()
+    amount = models.DecimalField(max_digits=8, decimal_places=2)
     date = models.DateField()
     user = models.ForeignKey(User, on_delete=models.CASCADE, null=True)
